Move mouse button constants into browser platform
