+++ web/src/lib/platform/browser/mouse-button.ts
